Authenticate verification code to the database entry (#22)
* Redirect to verify page on success

* Authenticate verification code with the database entry

When the user enters in their email address, a verification code is
sent to their email. The user then input the verification code to
the redirect page. Once the verification code has been verified
with the database, it will redirect the user to a different page.
Pair programmed with Sharon and Chris.

* Fix lint errors

---------

Co-authored-by: Jae Min Lee <jaemin.lee@gordon.edu>, Sharon Patta <sharonepatta@gmail.com>